+++ sdk/kusto/arm-kusto/src/api/databaseOperations/operations.ts
@@ -12,7 +12,7 @@ import {
   databaseInviteFollowerResultDeserializer,
 } from "../../models/models.js";
 import { expandUrlTemplate } from "../../static-helpers/urlTemplate.js";
-import type { DatabaseInviteFollowerOptionalParams } from "./options.js";
+import type { DatabaseOperationsInviteFollowerOptionalParams } from "./options.js";
 import type { StreamableMethod, PathUncheckedResponse } from "@azure-rest/core-client";
 import { createRestError, operationOptionsToRequestParameters } from "@azure-rest/core-client";
 
@@ -22,7 +22,7 @@ export function _inviteFollowerSend(
   clusterName: string,
   databaseName: string,
   parameters: DatabaseInviteFollowerRequest,
-  options: DatabaseInviteFollowerOptionalParams = { requestOptions: {} },
+  options: DatabaseOperationsInviteFollowerOptionalParams = { requestOptions: {} },
 ): StreamableMethod {
   const path = expandUrlTemplate(
     "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Kusto/clusters/{clusterName}/databases/{databaseName}/inviteFollower{?api%2Dversion}",
@@ -68,7 +68,7 @@ export async function inviteFollower(
   clusterName: string,
   databaseName: string,
   parameters: DatabaseInviteFollowerRequest,
-  options: DatabaseInviteFollowerOptionalParams = { requestOptions: {} },
+  options: DatabaseOperationsInviteFollowerOptionalParams = { requestOptions: {} },
 ): Promise<DatabaseInviteFollowerResult> {
   const result = await _inviteFollowerSend(
     context,