+++ sdk/kusto/arm-kusto/src/classic/databaseOperations/index.ts
@@ -2,39 +2,41 @@
 // Licensed under the MIT License.
 
 import type { KustoManagementContext } from "../../api/kustoManagementContext.js";
-import { inviteFollower } from "../../api/database/operations.js";
-import type { DatabaseInviteFollowerOptionalParams } from "../../api/database/options.js";
+import { inviteFollower } from "../../api/databaseOperations/operations.js";
+import type { DatabaseOperationsInviteFollowerOptionalParams } from "../../api/databaseOperations/options.js";
 import type {
   DatabaseInviteFollowerRequest,
   DatabaseInviteFollowerResult,
 } from "../../models/models.js";
 
-/** Interface representing a Database operations. */
-export interface DatabaseOperations {
+/** Interface representing a DatabaseOperations operations. */
+export interface DatabaseOperationsOperations {
   /** Generates an invitation token that allows attaching a follower database to this database. */
   inviteFollower: (
     resourceGroupName: string,
     clusterName: string,
     databaseName: string,
     parameters: DatabaseInviteFollowerRequest,
-    options?: DatabaseInviteFollowerOptionalParams,
+    options?: DatabaseOperationsInviteFollowerOptionalParams,
   ) => Promise<DatabaseInviteFollowerResult>;
 }
 
-function _getDatabase(context: KustoManagementContext) {
+function _getDatabaseOperations(context: KustoManagementContext) {
   return {
     inviteFollower: (
       resourceGroupName: string,
       clusterName: string,
       databaseName: string,
       parameters: DatabaseInviteFollowerRequest,
-      options?: DatabaseInviteFollowerOptionalParams,
+      options?: DatabaseOperationsInviteFollowerOptionalParams,
     ) => inviteFollower(context, resourceGroupName, clusterName, databaseName, parameters, options),
   };
 }
 
-export function _getDatabaseOperations(context: KustoManagementContext): DatabaseOperations {
+export function _getDatabaseOperationsOperations(
+  context: KustoManagementContext,
+): DatabaseOperationsOperations {
   return {
-    ..._getDatabase(context),
+    ..._getDatabaseOperations(context),
   };
 }